Add files via upload
This new push sets up the framework to give CCL locos names instead of their number id. All changes occurred within the pathAnimator js file. Added a new constant dictionary _modlocos contains their number id as keys and values are set up as their display names. A modification is done to interpretCarLine to check the name of the loco with the new method added called "interpretLocoType", then this name is used in place of the ID. Method interpretLocoType takes an input called name, this is the raw name from the CSV file, this may be used to determine the loco name based on the dictionary, alternatively, returned directly if it is one of the vanilla names.