Output Change
This commit modified Basereporter.cs in a way that permitted a new csv file to be created every time it starts logging, however, these files are now stored in /Sessions folder within the directory. There is also code to ensure that the folder exists and create it if otherwise. Error catching is also added.